test(sdk): split team runtime coverage
